+++ tests/update_handlers/tdct_test.py
@@ -5,6 +5,24 @@ from streaming_data_types.timestamps_tdct import deserialise_tdct
 from forwarder.update_handlers.tdct_serialiser import tdct_PVASerialiser
 
 
+def test_tdct_serialiser_handles_positive_relative_timestamps():
+    input_relative_timestamps = np.array([10, 11, 12]).astype(np.int32)
+    # This is the timestamp of the PV update
+    reference_timestamp = 10
+    serialiser = tdct_PVASerialiser("tst_source")
+
+    update = NTScalar("ai").wrap(input_relative_timestamps)
+    update.timeStamp.secondsPastEpoch = 0
+    update.timeStamp.nanoseconds = reference_timestamp
+
+    message = serialiser.serialise(update)
+
+    published_data = deserialise_tdct(message[0])
+    assert np.array_equal(
+        published_data.timestamps, input_relative_timestamps + reference_timestamp
+    ), "Expected the relative timestamps from the EPICS update to have been converted to unix timestamps"
+
+
 def test_tdct_serialiser_handles_negative_relative_timestamps():
     input_relative_timestamps = np.array([-3, -2, -1]).astype(np.int32)
     # This is the timestamp of the PV update
@@ -40,3 +58,21 @@ def test_tdct_publisher_publishes_successfully_when_there_is_only_a_single_chopp
         published_data.timestamps,
         np.atleast_1d(input_relative_timestamps + reference_timestamp),
     ), "Expected the relative timestamps from the EPICS update to have been converted to unix timestamps"
+
+
+def test_tdct_serialiser_handles_empty_array():
+    input_relative_timestamps = np.array([]).astype(np.int32)
+    # This is the timestamp of the PV update
+    reference_timestamp = 10
+    serialiser = tdct_PVASerialiser("tst_source")
+
+    update = NTScalar("ai").wrap(input_relative_timestamps)
+    update.timeStamp.secondsPastEpoch = 0
+    update.timeStamp.nanoseconds = reference_timestamp
+
+    message = serialiser.serialise(update)
+
+    assert message == (
+        None,
+        None,
+    ), "Expected serialiser to skip empty timestamp array from the EPICS update"